+++ frb_rust/src/lockable/mod.rs
@@ -1,5 +1,5 @@
 pub(crate) mod base;
-pub(crate) mod implementor;
 mod immediate_invalid_borrow_checker;
-mod lock_order_info;
-mod lockable_order;
+pub(crate) mod implementor;
+mod order;
+mod order_info;
